Add logging at validation failure in plan stage
The background here is to cover the case that the .terraform.lock.hcl
file has been accidentaly removed from git. Since a new .terraform.lock.hcl
will have been generated by terraform init in that case, it will show up
as an untracked file in the output of git status.